requirements.txt
datasets>=2.20.0
python-dotenv>=1.0.1
streamlit>=1.36.0
openai>=1.0.0

# Testing
pytest>=8.0.0

